+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.0.0-alpha1 [16th of June 2018]
 - Introduce data helper classes
 - Update NPM dependencies for theme
+- Renamed class methods to follow camelCase 
 
 1.0.0-alpha [15th of June 2018]
 - initial release